fixed some shit.

+++ database/migrations/2014_10_12_000000_create_users_table.php
@@ -16,7 +16,7 @@ class CreateUsersTable extends Migration
         Schema::create('users', function (Blueprint $table) {
             $table->increments('id');
             $table->string('name');
-            $table->decimal('budgetStartDay');
+            $table->integer('budgetStartDay');
             $table->string('email')->unique();
             $table->string('password');
             $table->rememberToken();
